+++ src/App.jsx
@@ -1,4 +1,5 @@
 import "./App.css";
+import Day14Throttling from "./components/Day14Throttling/Day14Throttling";
 
 function App() {
   // function throttle(func, delay) { // ? using timeoutId
@@ -13,31 +14,9 @@ function App() {
   //   };
   // }
 
-  function throttle(func, delay) {
-    // ? using locked variable
-    let locked = false;
-    return () => {
-      if (!locked) {
-        func(); // ? api call
-        locked = true;
-        setTimeout(() => {
-          locked = false;
-        }, delay);
-      }
-    };
-  }
-
   return (
-    <main className="m-4 flex flex-row items-center gap-5">
-      <button
-        className="bg-blue-500 px-5 py-1 rounded-md"
-        onClick={throttle(() => {
-          let timeClicked = new Date();
-          console.log(timeClicked);
-        }, 1000)}
-      >
-        Search
-      </button>
+    <main>
+      <Day14Throttling />
     </main>
   );
 }
